small cleanups + setting up new dirs + Added chakra provider
