+++ src/App.jsx
@@ -1,7 +1,8 @@
 import React from 'react';
 import { Route, Routes } from 'react-router-dom';
 import { CookiesProvider } from 'react-cookie';
-import Home from './pages/Home/Home';
+import { ChakraProvider } from '@chakra-ui/react';
+// import Home from './pages/Home/Home';
 import ForgotPassword from './pages/ForgotPassword/ForgotPassword';
 import Dogs from './pages/Dogs/Dogs';
 import AddDog from './pages/AddDog/AddDog';
@@ -10,20 +11,23 @@ import ErrorPage from './pages/ErrorPage';
 import Navbar from './components/Navbar';
 import './App.css';
 import './common/global.css';
+import LoginPage from './pages/Login/LoginPage';
 
 const App = () => {
   return (
-    <CookiesProvider>
-      <Navbar />
-      <Routes>
-        <Route path="/login" element={<Home />} />
-        <Route path="/forgot-password" element={<ForgotPassword />} />
-        <Route path="/" element={<Dogs />} />
-        <Route path="/dogs/new" element={<AddDog />} />
-        <Route path="/facilities" element={<Facilities />} />
-        <Route path="*" element={<ErrorPage />} />
-      </Routes>
-    </CookiesProvider>
+    <ChakraProvider>
+      <CookiesProvider>
+        <Navbar />
+        <Routes>
+          <Route path="/login" element={<LoginPage />} />
+          <Route path="/forgot-password" element={<ForgotPassword />} />
+          <Route path="/" element={<Dogs />} />
+          <Route path="/dogs/new" element={<AddDog />} />
+          <Route path="/facilities" element={<Facilities />} />
+          <Route path="*" element={<ErrorPage />} />
+        </Routes>
+      </CookiesProvider>
+    </ChakraProvider>
   );
 };
